Clean up the match_parse extension
Also, don't use an extra directory for that,
and put the test in the standard test suite.
