--- twill/extensions/match_parse/test-match_parse.py
@@ -1,1 +1,0 @@
-#
